requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0
google-cloud-secret-manager==2.16.1
google-auth==2.22.0
Pillow==10.0.0
numpy>=2.2.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
werkzeug==2.3.7
